Refactor PostCSS configuration: replace validate-var plugin with validate-variables and update paths
