+++ packages/survey-creator-core/postcss/postcss.config.cjs
@@ -2,14 +2,17 @@
 /* eslint-env node */
 /* eslint-disable @typescript-eslint/no-var-requires */
 
+const path = require("path");
+
+// Resolve paths relative to the package root (postcss-cli runs with that as cwd).
+const packageRoot = path.resolve(__dirname, "..");
+
 module.exports = {
   plugins: [
     require("postcss-import"),
-    require("postcss-calc"),
-    require("autoprefixer"),
-    require("./postcss-validate-var")({
+    require("./validate-variables")({
       extraFiles: [
-        "node_modules/survey-core/survey-core.css"
+        path.join(packageRoot, "node_modules/survey-core/survey-core.css")
       ],
       // Custom properties that are set at runtime via JS (element.style.setProperty)
       // and therefore never appear as `--name:` declarations in any CSS file.
@@ -17,6 +20,8 @@ module.exports = {
         "--animation-width",
       ]
     }),
+    require("postcss-calc"),
+    require("autoprefixer"),
     require("postcss-fail-on-warn")
   ]
 };
