+++ packages/survey-creator-core/postcss/validate-variables.js
